Change precision to f64

+++ audio-engine-instrument-bowed-string/src/bow.rs
@@ -1,9 +1,9 @@
 #[derive(Debug, Default, Copy, Clone)]
 pub struct Bow {
     /// Bow Pressure (fb)
-    pub pressure: f32,
+    pub pressure: f64,
     /// Bow Speed in meters per second (vb/bowVel)
-    pub velocity: f32,
+    pub velocity: f64,
 }
 
-pub const BOW_FREE_PARAMETER: f32 = 100.0;
+pub const BOW_FREE_PARAMETER: f64 = 100.0;
